nemotron/requirements.txt
langchain>=0.2.11
langchain-community>=0.2.11
langchain-nvidia-ai-endpoints>=0.2.0
chromadb>=0.5.0
pypdf>=4.2.0
python-dotenv>=1.0.0
requests>=2.32.0
